refactor: use env var for @timing decorator
